Add files via upload
The file WorldBank.txt contains the total fertility rate and the GDP per capita in US$ of 180 countries and regions in 2013 collected by the World Bank. The world-stats program reads the dataset and summarize the dataset using the statistical measures such as the mean, the median, the variance, and the standard deviation of a variable in a data set.